fix: Enforce behaviour
